% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble.R
\name{as_tibble.palettes_colour}
\alias{as_tibble.palettes_colour}
\alias{as_tibble.palettes_palette}
\title{Convert colour palette to a tibble}
\usage{
\method{as_tibble}{palettes_colour}(x, ...)

\method{as_tibble}{palettes_palette}(x, ...)
}
\arguments{
\item{x}{An object of class \code{palettes_palette} or \code{palettes_colour}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Function converts a colour palette object to a tibble.
}
\examples{
x <- pal_colour(c("red", "green", "blue"))
as_tibble(x)

y <- pal_palette(
  pal1 = c("red", "green", "blue"),
  pal2 = c("yellow", "orange", "purple")
)
as_tibble(y)
}
