% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-palettes.R
\docType{data}
\name{carto_palettes}
\alias{carto_palettes}
\alias{discrete_palettes,}
\alias{sequential_palettes,}
\alias{diverging_palettes}
\alias{discrete_palettes}
\alias{sequential_palettes}
\title{Cartography palettes}
\format{
\subsection{\code{discrete_palettes}}{

An object of class \code{palettes_palette} with 6
colour palettes. Use \code{names(discrete_palettes)} to return all palette names.
}

\subsection{\code{sequential_palettes}}{

An object of class \code{palettes_palette} with 19
colour palettes. Use \code{names(sequential_palettes)} to return all palette names.
}

\subsection{\code{diverging_palettes}}{

An object of class \code{palettes_palette} with 7
colour palettes. Use \code{names(diverging_palettes)} to return all palette names.
}
}
\source{
\url{https://github.com/CartoDB/CartoColor/}
}
\usage{
discrete_palettes

sequential_palettes

diverging_palettes
}
\description{
Discrete, sequential, and diverging palettes created by cartographers.
}
\examples{
# Get all palettes by name.
names(discrete_palettes)
names(sequential_palettes)
names(diverging_palettes)

# Plot all palettes.
plot(discrete_palettes)
plot(sequential_palettes)
plot(diverging_palettes)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}
}
\keyword{datasets}
