% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{colour-mixing-arithmetic}
\alias{colour-mixing-arithmetic}
\alias{color-mixing-arithmetic}
\alias{+.palettes_colour}
\title{Mix colour vectors with arithmetic operators}
\usage{
\method{+}{palettes_colour}(e1, e2)
}
\arguments{
\item{e1, e2}{\link{pal_colour} vectors or objects which can be coerced to
them.}
}
\value{
Binary \code{+} returns a \link{pal_colour} vector.
}
\description{
These binary operators perform "arithmetic" on \link{pal_colour} vectors
(or objects which can be coerced to them).
}
\examples{
x <- pal_colour("red")
y <- pal_colour("blue")
x + y
}
