% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{colour-mixing-math}
\alias{colour-mixing-math}
\alias{color-mixing-math}
\alias{sum.palettes_colour}
\alias{cumsum.palettes_colour}
\title{Mix colour vectors with math functions}
\usage{
\method{sum}{palettes_colour}(..., na.rm = FALSE)

\method{cumsum}{palettes_colour}(x)
}
\arguments{
\item{...}{Colour vectors of class \code{palettes_colour}.}

\item{na.rm}{Whether to include missing values. Either \code{TRUE} or \code{FALSE}.}

\item{x}{An object of class \code{palettes_colour}.}
}
\value{
A \link{pal_colour} vector.
}
\description{
These functions perform "math" on \link{pal_colour} vectors
(or objects which can be coerced to them).
}
\examples{
x <- pal_colour(c("red", "blue"))
sum(x)

x <- pal_colour(c("red", "blue", "yellow"))
cumsum(x)
}
