% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{list_colour_symbols}
\alias{list_colour_symbols}
\alias{list_color_symbols}
\title{Symbols to use in colour previews}
\usage{
list_colour_symbols()
}
\value{
This function is called for its side effects and has no return value.
}
\description{
List the symbols available to use in colour previews.
}
\details{
By default, Unicode characters are used for symbols in colour previews in
UTF-8 supported outputs. They automatically fall back to ASCII characters
when the output does not support them.

To change the symbol used for colour previews, set the \code{palettes.print_symbol}
option to a symbol name listed in \code{list_colour_symbols()}.
}
\examples{
list_colour_symbols()
}
\seealso{
\code{\link[=palettes-options]{help("palettes-options")}}, \code{\link[cli:is_utf8_output]{cli::is_utf8_output()}}
}
