% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math}
\alias{math}
\alias{sum}
\alias{cumsum}
\title{Math functions}
\usage{
sum(x)
cumsum(x)
}
\arguments{
\item{x}{A \link{pal_colour} vector or an object which can be coerced to one.}
}
\value{
A \link{pal_colour} vector.
}
\description{
These functions perform "math" on \link{pal_colour} vectors
(or objects which can be coerced to them).
}
\examples{
x <- pal_colour(c("red", "blue"))
sum(x)

x <- pal_colour(c("red", "blue", "yellow"))
cumsum(x)
}
