% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_blend.R
\name{pal_blend}
\alias{pal_blend}
\title{#' Colour blending
#'
#' Blends two colors using RGB channel-wise blend functions. Valid blend modes
#' are multiply, darken, lighten, screen, overlay, burn, and dodge.
#'
#' @param a,b Objects of class \code{palettes_colour}.
#' @param mode The colour blending mode.
#'
#' @return An object of class \code{palettes_colour}.
#' @export}
\usage{
pal_blend(a, b, mode = blend_modes())
}
\description{
#' Colour blending
#'
#' Blends two colors using RGB channel-wise blend functions. Valid blend modes
#' are multiply, darken, lighten, screen, overlay, burn, and dodge.
#'
#' @param a,b Objects of class \code{palettes_colour}.
#' @param mode The colour blending mode.
#'
#' @return An object of class \code{palettes_colour}.
#' @export
}
