% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{pal_colour}
\alias{pal_colour}
\alias{is_colour}
\alias{as_colour}
\title{\code{colour} vector}
\usage{
pal_colour(x = character())

is_colour(x)

as_colour(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{pal_colour()}: A character vector of any of the three kinds of R colour
specifications.
\item For \code{as_colour()}: An object to be coerced.
\item For \code{is_colour()}: An object to test.
}}
}
\value{
An S3 vector of class \code{palettes_colour}.
}
\description{
This creates a character vector that represents colours so when it is
printed, colours will be formatted as hexadecimal strings.
}
\details{
Colours can be specified using either:
\itemize{
\item A colour name (as listed by colours())
\item a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb)
\item A positive integer i meaning \code{palette()[i]}.
}
}
\examples{
pal_colour(c("red", "#00FF00", "blue"))
is_colour(c("red", "#00FF00", "blue"))
as_colour("white")
}
