% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_invert.R
\name{pal_invert.palettes_palette}
\alias{pal_invert.palettes_palette}
\title{#' @export
#' @rdname pal_invert}
\usage{
\method{pal_invert}{palettes_palette}(x)
}
\description{
#' @export
#' @rdname pal_invert
}
