% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_invert.R
\name{pal_invert}
\alias{pal_invert}
\title{#' Invert a colour vector
#'
#' @param x An object of class \code{palettes_palette} or \code{palettes_colour}.
#'
#' @return An object of the same type as \code{x}.
#' @export
#'
#' @examples
#' pal_invert("white")}
\usage{
pal_invert(x)
}
\description{
#' Invert a colour vector
#'
#' @param x An object of class \code{palettes_palette} or \code{palettes_colour}.
#'
#' @return An object of the same type as \code{x}.
#' @export
#'
#' @examples
#' pal_invert("white")
}
