% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{pal_palette}
\alias{pal_palette}
\alias{is_palette}
\alias{as_palette}
\title{Colour palettes}
\usage{
pal_palette(...)

is_palette(x)

as_palette(x)
}
\arguments{
\item{...}{\itemize{
\item For \code{pal_palette()}: A named list of character vectors of any of the three
kinds of R colour specifications, or a named list of colour vectors of
class \code{palettes_colour}.
}}

\item{x}{\itemize{
\item For \code{as_palette()}: An object to be coerced.
\item For \code{is_palette()}: An object to test.
}}
}
\value{
An S3 list of class \code{palettes_palette}.
}
\description{
This creates a list of colour vectors.
}
\details{
Colours can be specified using either:
\itemize{
\item Hexadecimal strings of the form \code{"#RRGGBB"} or \code{"#RRGGBBAA"}
\item Colour names from \code{grDevices::colors()}
\item Positive integers \code{i} that index into \code{grDevices::palette()[i]}
}
}
\examples{
pal_palette(
  pal1 = c("red", "green", "blue"),
  pal2 = c("yellow", "orange", "purple")
)

x <- list(
  pal1 = c("red", "green", "blue"),
  pal2 = c("yellow", "orange", "purple")
)
as_palette(x)
}
\seealso{
\code{pal_colour()}
}
