% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{palettes-options}
\alias{palettes-options}
\title{Package options}
\description{
Options that adjust the behaviour of the palettes package.
}
\details{
These options can be set via \code{\link[=options]{options()}} and queried via \code{\link[=getOption]{getOption()}}.
}
\section{Options for the palettes package}{


\describe{

\item{\code{palettes.symbol}:}{

Character string setting the symbol used for colour previews. See
\code{cli::list_symbols()} for a list of symbol choices. Defaults to \code{"bullet"}.

}

}
}

