% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-package.R
\docType{package}
\name{palettes-package}
\alias{palettes}
\alias{palettes-package}
\title{palettes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael McCarthy \email{first.last@example.com}

}
\keyword{internal}
