% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-palettes.R
\docType{data}
\name{penguin_palettes}
\alias{penguin_palettes}
\title{Palmer penguins palettes}
\format{
\subsection{\code{penguin_palettes}}{

An object of class \code{palettes_palette} with 6
colour palettes. Use \code{names(penguin_palettes)} to return all palette names.
}
}
\source{
\url{https://github.com/allisonhorst/palmerpenguins/}
}
\usage{
penguin_palettes
}
\description{
Palettes inspired by the Palmer penguins.
}
\examples{
# Get all palettes by name.
names(penguin_palettes)

# Plot all palettes.
plot(penguin_palettes)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}
}
\author{
\href{https://github.com/allisonhorst}{Allison Horst}
}
\keyword{datasets}
