% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{performance_palettes}
\alias{performance_palettes}
\title{Performance palettes}
\format{
\subsection{\code{performance_palettes}}{

An object of class \code{palettes_palette} with 1
colour palettes. All colours in each palette are distinguishable with
deuteranopia, protanopia, and tritanopia. Use \code{names(performance_palettes)}
to return all palette names.
}
}
\source{
\url{https://github.com/easystats/performance/}
}
\usage{
performance_palettes
}
\description{
A colourblind accessible palette for visualizing performance.
}
\examples{
# Get all palettes by name.
names(performance_palettes)

# Plot all palettes.
plot(performance_palettes)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}
}
\keyword{datasets}
