% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pnw_palettes}
\alias{pnw_palettes}
\title{Pacific Northwest palettes}
\format{
\subsection{\code{pnw_palettes}}{

An object of class \code{palettes_palette} with 14 colour
palettes. Use \code{names(pnw_palettes)} to return all palette names.
}
}
\source{
\url{https://github.com/jakelawlor/PNWColors}
}
\usage{
pnw_palettes
}
\description{
Palettes inspired by Jake Lawlor's photos of the dreamiest, most colourful,
PNW-iest places in Washington State.
}
\examples{
# Get all palettes by name.
names(pnw_palettes)

# Plot all palettes.
plot(pnw_palettes)
}
\seealso{
\code{\link[=pal_palette]{pal_palette()}}, \code{\link[=pal_colour]{pal_colour()}}, \code{\link[PNWColors:pnw_palette]{PNWColors::pnw_palette()}}
}
\author{
\href{https://github.com/jakelawlor}{Jake Lawlor}
}
\keyword{datasets}
