% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{arithmetic}
\alias{arithmetic}
\alias{+}
\title{Arithmetic operators}
\usage{
x + y
}
\arguments{
\item{x, y}{\link{pal_colour} vectors or objects which can be coerced to
them.}
}
\value{
Unary \code{+} returns a \link{pal_colour} vector.
}
\description{
These unary operators perform "arithmetic" on \link{pal_colour} vectors
(or objects which can be coerced to them).
}
\examples{
x <- pal_colour("red")
y <- pal_colour("blue")
x + y
}
