% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-arithmetic.R
\name{colour-mixing-arithmetic}
\alias{colour-mixing-arithmetic}
\alias{color-mixing-arithmetic}
\alias{+.palettes_colour}
\title{Mix colour vectors with arithmetic operators}
\usage{
\method{+}{palettes_colour}(e1, e2)
}
\arguments{
\item{e1, e2}{Colour vectors of class \code{palettes_colour}.}
}
\value{
The binary operators return colour vectors of class \code{palettes_colour} containing
the result of the element by element operations. If involving a zero-length
vector the result has length zero. Otherwise, the elements of shorter vectors
are recycled as necessary. The \code{+} operator is for additive colour mixing.
}
\description{
These binary operators mix colour vectors with arithmetic operators.
}
\examples{
x <- pal_colour("red")
y <- pal_colour("blue")
x + y
}
