% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{pal_palette}
\alias{pal_palette}
\title{\code{color} palettes}
\usage{
pal_palette(...)
}
\arguments{
\item{...}{\itemize{
\item For \code{pal_palette()}: A list of character vector of any of the three kinds of
R colour specifications.
}}
}
\value{
An S3 list of class \code{palettes_palette}.
}
\description{
This creates a list of color vectors.
}
\details{
Colours can be specified using either:
\itemize{
\item A color name (as listed by colors())
\item a hexadecimal string of the form "#rrggbb" or "#rrggbbaa" (see rgb)
\item A positive integer i meaning \code{palette()[i]}.
}
}
\examples{
pal_palette(
  pal1 = c("red", "green", "blue"),
  pal2 = c("yellow", "orange", "purple")
)
}
\seealso{
\code{pal_color()}
}
