% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes-package.R
\docType{package}
\name{palettes-package}
\alias{palettes}
\alias{palettes-package}
\title{palettes: Methods for Colours}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mccarthy-m-g.github.io/palettes/}
}

}
\author{
\strong{Maintainer}: Michael McCarthy \email{m.mccarthy1624@gmail.com}

}
\keyword{internal}
