% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.palettes_colour}
\alias{plot.palettes_colour}
\alias{plot.palettes_palette}
\title{Plot colour palette}
\usage{
\method{plot}{palettes_colour}(x, ...)

\method{plot}{palettes_palette}(x, ...)
}
\arguments{
\item{x}{An object of class \code{palettes_palette} or \code{palettes_colour}.}
}
\description{
Plots a colour palette object.
}
\examples{
x <- pal_colour(c("red", "green", "blue"))
plot(x)

y <- pal_palette(my_pal = c("red", "green", "blue"))
plot(y)

z <- pal_palette(
  pal1 = c("red", "green", "blue"),
  pal2 = c("yellow", "orange", "purple")
)
plot(z)
}
